requirements.txt
# Health Metrics LLM Prototype Requirements
# Python 3.10+

# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Data processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
pyarrow>=14.0.0,<16.0.0

# LLM/AI
openai>=1.0.0,<2.0.0

# Azure Storage
azure-storage-blob>=12.19.0,<13.0.0

# Redis Cache
redis>=5.0.0,<6.0.0

# Web Framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0

# Jupyter notebook support
ipykernel>=6.25.0,<8.0.0
